requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-Cors==4.0.0
gunicorn==21.2.0
Werkzeug==3.1.4

# Database
pymongo==4.6.1

# Google APIs
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.0

# Authentication
svix==1.15.0

# Image/PDF generation for certificates
pillow==12.0.0
reportlab==4.4.6

# Utilities
python-dotenv==1.0.0
requests==2.31.0
